[Refactor] - Change from overrides to extends (#2)
* Update README

* Add french translations

* Extend Unpublish Initiative Admin Command

* Add Makefile for extends

* Extend Decidim Initiatives Permissions

* Extend Decidim Initiatives Admin Permissions

* Extend Decidim Initiatives Controller

* Extend Decidim Initiatives Admin Controller

* Extend Decidim Comments Comment model

* Add comments in decidim initiatives model

* Add Make command for overrides

* Extend Decidim Initiatives AdminLog Presenter

* Fix Extends loading

* Add Makefile commands in README
